refactor: TestConstant 클래스명 의미에 맞게 변경
