+++ src/test/java/piece/ElephantTest.java
@@ -3,21 +3,21 @@ package piece;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static piece.Country.Cho;
-import static testutil.StaticTest.B3;
-import static testutil.StaticTest.B7;
-import static testutil.StaticTest.C2;
-import static testutil.StaticTest.C8;
-import static testutil.StaticTest.D5;
-import static testutil.StaticTest.E4;
-import static testutil.StaticTest.E5;
-import static testutil.StaticTest.E6;
-import static testutil.StaticTest.F5;
-import static testutil.StaticTest.F6;
-import static testutil.StaticTest.G2;
-import static testutil.StaticTest.G6;
-import static testutil.StaticTest.G8;
-import static testutil.StaticTest.H3;
-import static testutil.StaticTest.H7;
+import static testutil.TestConstant.B3;
+import static testutil.TestConstant.B7;
+import static testutil.TestConstant.C2;
+import static testutil.TestConstant.C8;
+import static testutil.TestConstant.D5;
+import static testutil.TestConstant.E4;
+import static testutil.TestConstant.E5;
+import static testutil.TestConstant.E6;
+import static testutil.TestConstant.F5;
+import static testutil.TestConstant.F6;
+import static testutil.TestConstant.G2;
+import static testutil.TestConstant.G6;
+import static testutil.TestConstant.G8;
+import static testutil.TestConstant.H3;
+import static testutil.TestConstant.H7;
 
 import game.Board;
 import java.util.Map;
