+++ src/test/java/piece/GeneralTest.java
@@ -2,16 +2,16 @@ package piece;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static testutil.StaticTest.A1;
-import static testutil.StaticTest.A2;
-import static testutil.StaticTest.A3;
-import static testutil.StaticTest.B1;
-import static testutil.StaticTest.D1;
-import static testutil.StaticTest.D5;
-import static testutil.StaticTest.E4;
-import static testutil.StaticTest.E5;
-import static testutil.StaticTest.E6;
-import static testutil.StaticTest.F5;
+import static testutil.TestConstant.A1;
+import static testutil.TestConstant.A2;
+import static testutil.TestConstant.A3;
+import static testutil.TestConstant.B1;
+import static testutil.TestConstant.D1;
+import static testutil.TestConstant.D5;
+import static testutil.TestConstant.E4;
+import static testutil.TestConstant.E5;
+import static testutil.TestConstant.E6;
+import static testutil.TestConstant.F5;
 
 import game.Board;
 import java.util.Map;
