+++ src/test/java/piece/HorseTest.java
@@ -3,20 +3,20 @@ package piece;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static piece.Country.Cho;
-import static testutil.StaticTest.C4;
-import static testutil.StaticTest.C6;
-import static testutil.StaticTest.D3;
-import static testutil.StaticTest.D5;
-import static testutil.StaticTest.D7;
-import static testutil.StaticTest.E4;
-import static testutil.StaticTest.E5;
-import static testutil.StaticTest.E6;
-import static testutil.StaticTest.F3;
-import static testutil.StaticTest.F5;
-import static testutil.StaticTest.F6;
-import static testutil.StaticTest.F7;
-import static testutil.StaticTest.G4;
-import static testutil.StaticTest.G6;
+import static testutil.TestConstant.C4;
+import static testutil.TestConstant.C6;
+import static testutil.TestConstant.D3;
+import static testutil.TestConstant.D5;
+import static testutil.TestConstant.D7;
+import static testutil.TestConstant.E4;
+import static testutil.TestConstant.E5;
+import static testutil.TestConstant.E6;
+import static testutil.TestConstant.F3;
+import static testutil.TestConstant.F5;
+import static testutil.TestConstant.F6;
+import static testutil.TestConstant.F7;
+import static testutil.TestConstant.G4;
+import static testutil.TestConstant.G6;
 
 import game.Board;
 import java.util.Map;
