+++ src/test/java/piece/RookTest.java
@@ -3,18 +3,18 @@ package piece;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static piece.Country.Cho;
-import static testutil.StaticTest.A5;
-import static testutil.StaticTest.B3;
-import static testutil.StaticTest.C5;
-import static testutil.StaticTest.D4;
-import static testutil.StaticTest.E1;
-import static testutil.StaticTest.E4;
-import static testutil.StaticTest.E5;
-import static testutil.StaticTest.E6;
-import static testutil.StaticTest.E9;
-import static testutil.StaticTest.F6;
-import static testutil.StaticTest.G5;
-import static testutil.StaticTest.I5;
+import static testutil.TestConstant.A5;
+import static testutil.TestConstant.B3;
+import static testutil.TestConstant.C5;
+import static testutil.TestConstant.D4;
+import static testutil.TestConstant.E1;
+import static testutil.TestConstant.E4;
+import static testutil.TestConstant.E5;
+import static testutil.TestConstant.E6;
+import static testutil.TestConstant.E9;
+import static testutil.TestConstant.F6;
+import static testutil.TestConstant.G5;
+import static testutil.TestConstant.I5;
 
 import game.Board;
 import java.util.Map;
